Check if miniforge is already installed before re-downloading it and reinstalling it.

+++ run_onchange_20_setup_gui.sh
@@ -22,12 +22,13 @@ dconf load /org/cinnamon/desktop/peripherals/keyboard/ <~/.local/share/chezmoi/k
 
 #############################
 # MINIFORGE LATEST
-wget -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
-chmod u+x ./Miniforge3.sh
-./Miniforge3.sh -b
-~/miniforge3/condabin/conda init
-rm Miniforge3.sh
-
+if [ ! -d "$HOME/miniforge3" ]; then
+  wget -q -O Miniforge3.sh "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
+  chmod u+x ./Miniforge3.sh
+  ./Miniforge3.sh -b
+  ~/miniforge3/condabin/conda init
+  rm Miniforge3.sh
+fi
 # DISCORD LATEST
 #wget -q -O discord.deb https://discord.com/api/download?platform=linux&format=deb
 #sudo dpkg -i discord.deb
